Limit that only changesets lower than HEAD can be deleted
- Add new GetHeadVersion QR which fetches the max version from the space table
- Change DeleteChangesets QR to throw an ILLEGAL_ARGUMENT error if the user tries to specify a value larger than HEAD (uses new GetHeadVersion QR)
- Add test deleteChangesetsLargerThanHeadValue() which tests that the above error is thrown
- Also refactor all occurrences of "version" to be long rather than int

+++ xyz-hub-service/src/main/java/com/here/xyz/hub/task/FeatureTaskHandler.java
@@ -162,7 +162,7 @@ public class FeatureTaskHandler {
    * If the value exists for a space and it points to a value > 0, that is the version of the latest write to that space as it has been
    * performed by this service-node.
    */
-  private static ConcurrentHashMap<String, Integer> latestSeenContentVersions = new ConcurrentHashMap<>();
+  private static ConcurrentHashMap<String, Long> latestSeenContentVersions = new ConcurrentHashMap<>();
 
   /**
    * Contains the amount of all in-flight requests for each storage ID.
@@ -512,9 +512,9 @@ public class FeatureTaskHandler {
     }
   }
 
-  static void setLatestSeenContentVersion(Space space, int version) {
+  static void setLatestSeenContentVersion(Space space, long version) {
     if ((space.isEnableHistory() || space.isEnableGlobalVersioning()) && version > 0)
-      latestSeenContentVersions.compute(space.getId(), (spaceId, currentVersion) -> Math.max(currentVersion != null ? currentVersion : 0,
+      latestSeenContentVersions.compute(space.getId(), (spaceId, currentVersion) -> Math.max(currentVersion != null ? currentVersion : 0L,
           version));
   }
 
@@ -538,7 +538,7 @@ public class FeatureTaskHandler {
       long timerId = Service.vertx.setTimer(interval, tId -> {
         timerMap.remove(nc.space.getId());
         ContentModifiedNotification cmn = new ContentModifiedNotification().withSpace(nc.space.getId());
-        Integer spaceVersion = latestSeenContentVersions.get(nc.space.getId());
+        Long spaceVersion = latestSeenContentVersions.get(nc.space.getId());
         if (spaceVersion != null) cmn.setSpaceVersion(spaceVersion);
         if (adminNotification) {
           //Send it to the modification SNS topic
